requirements.txt
einops==0.8.1
flash_attn==2.6.3
numpy==1.23.0
Pillow==11.3.0
timm==0.6.13
torch==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.5
transformers==4.46.2
